backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0
textstat==0.7.3
groq==0.4.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.7
aiohttp==3.9.3
fake-useragent==1.4.0
